+++ Avisynth3/trunk/avisynth/src/core/clip.h
@@ -66,11 +66,11 @@ public:  //clip general interface
  
   //get the frame n
   //undefined behavior if frame n don't exist
-  virtual CPVideoFrame GetFrame(int n) const = 0;
+  virtual CPVideoFrame GetFrame(long n) const = 0;
 
   //fill the passed buffer with audio samples, and returns position after them
   //out of bounds values are allowed, the excess is filled with blank noise
-  virtual BYTE * GetAudio(BYTE * buffer, long long start, int count) const = 0;  
+  virtual BYTE * GetAudio(BYTE * buffer, long long start, long count) const = 0;  
 
 
 public:  //filter chain simplication methods
