+++ Avisynth3/trunk/avisynth/src/filters/audio/killvideo.h
@@ -55,8 +55,8 @@ public:  //clip general interface
   virtual PEnvironment const& GetEnvironment() const { return GetChildEnvironment(); }
   virtual CPVideoInfo GetVideoInfo() const;
 
-  virtual CPVideoFrame GetFrame(int n) const;
-  virtual BYTE * GetAudio(BYTE * buffer, long long start, int count) const { return GetChildAudio(buffer, start, count); }
+  virtual CPVideoFrame GetFrame(long n) const;
+  virtual BYTE * GetAudio(BYTE * buffer, long long start, long count) const { return GetChildAudio(buffer, start, count); }
 
 
 public:  //child changing clone
